% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{RHD-package}
\alias{RHD-package}
\title{Regularized halfspace depth for functional data}
\description{
The package provides tools for ranking functional data from centered-outward through the regularized halfspace depth (RHD).
\code{RHD} provides depth values and the associated rankings of functional data from the RHD and the outliers based on the outlier detection method linked to the RHD.
\code{AFSG} chooses the adjustment factor based simulated Gaussian processes for outlier detection method linked to the RHD.
}
\references{
Yeon, H., Dai, X., and Lopez-Pintado, S. (2022+). Regularized halfspace depth for functional data. In progress
}
\seealso{
\code{\link{RHD}}
\code{\link{AFSG}}
}
\author{
Hyemin Yeon \email{luckyhm@iastate.edu}
}
